+++ CHANGELOG.md
@@ -13,6 +13,7 @@ N/A
 #### Enhancements
 
 - Allow imbricated `AnimationType` options, `Array` and operator `+` to create `compound` animation by string. [#611](https://github.com/IBAnimatable/IBAnimatable/pull/611) by [@phimage](https://github.com/phimage)
+- Implement `Hashable` with `Hasher`. [#616](https://github.com/IBAnimatable/IBAnimatable/pull/616) by [@phimage](https://github.com/phimage)
 
 #### Bugfixes