+++ CHANGELOG.md
@@ -18,7 +18,7 @@ N/A
 
 #### Bugfixes
 
-N/A
+- Fix `swift build` [#623](https://github.com/IBAnimatable/IBAnimatable/pull/623) by [@phimage](https://github.com/phimage)
 
 ### [6.0.0](https://github.com/IBAnimatable/IBAnimatable/releases/tag/6.0.0)