+++ CHANGELOG.md
@@ -12,6 +12,20 @@ N/A
 
 #### Enhancements
 
+N/A
+
+#### Bugfixes
+
+N/A
+
+### [6.1.0](https://github.com/IBAnimatable/IBAnimatable/releases/tag/6.1.0)
+
+#### API breaking changes
+
+N/A
+
+#### Enhancements
+
 - Allow imbricated `AnimationType` options, `Array` and operator `+` to create `compound` animation by string. [#611](https://github.com/IBAnimatable/IBAnimatable/pull/611) by [@phimage](https://github.com/phimage)
 - Implement `Hashable` with `Hasher`. [#616](https://github.com/IBAnimatable/IBAnimatable/pull/616) by [@phimage](https://github.com/phimage)
 - Add `roundedPolygon`, a polygon `MaskType` with corner radius. [#619](https://github.com/IBAnimatable/IBAnimatable/pull/619) by [@phimage](https://github.com/phimage)